Move to modules/s3db/building.py
