+++ views/building/index.html
@@ -11,7 +11,7 @@ NZSEE {{=T("Safety Assessment Form")}} ({{=T("ATC-20 Rapid Evaluation modified f
 <li><a href='{{=URL(f="nzseel1", args=["create"])}}'>{{=T("Submit new Level 1 assessment (full form)")}}</a></li>
 <li><a href='{{=URL(f="nzseel2", args=["create"])}}'>{{=T("Submit new Level 2 assessment")}}</a></li>
 <li><a href='{{=URL(f="nzseel1")}}'>{{=T("List Level 1 assessments")}}</a></li>
-<li><a href='{{=URL(f="'nzseel2")}}'>{{=T("List Level 2 assessments")}}</a></li>
+<li><a href='{{=URL(f="nzseel2")}}'>{{=T("List Level 2 assessments")}}</a></li>
 </ul>
 <h3>{{=T("Reports")}}:</h3>
 <li><a href='{{=URL(f="report")}}'>{{=T("Snapshot Report")}}</a></li>
